Added @JvmStatic to verify methods and moved them to ContractUpgradeFlow.
